add skeleton files for usernetes + terraform on aws
This adds a base ubuntu image (ami) and next steps are to
debug how to get the instance ip addresses and hostnames
(in setup.sh) and write into /etc/hosts, or something
simple like that so they can be seen. Then we will want to
install usernetes custom on each of the control plane
and worker nodes, generate the join-command on the
control plane to share with workers, join, and then
test kubectl and the hack test app.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>