add slight more work to figure out the hostnames
It is still a bit of a hack. Ideally we can add identifiable hostnames
to the /etc/hosts of each. I am having trouble debugging because
there seems to be some delay in the info being available with the aws
Python sdk, and then I am not able to always shell into instances with
my PEM

Signed-off-by: vsoch <vsoch@users.noreply.github.com>